+++ pyogrio/tests/test_raw_io.py
@@ -6,6 +6,7 @@ import sys
 import numpy as np
 from numpy import array_equal
 import pytest
+import shapely
 
 from pyogrio import (
     list_layers,
@@ -583,17 +584,10 @@ def assert_equal_result(result1, result2):
     assert np.array_equal(meta1["fields"], meta2["fields"])
     assert np.array_equal(index1, index2)
     # a plain `assert np.array_equal(geometry1, geometry2)` doesn't work because
-    # the WKB values are not exactly equal, therefore parsing with pygeos to compare
+    # the WKB values are not exactly equal, therefore parsing with shapely to compare
     # with tolerance
-    try:
-        from shapely import from_wkb, equals_exact
-    except ImportError:
-        try:
-            from pygeos import from_wkb, equals_exact
-        except ImportError:
-            pytest.skip("Test requires pygeos or shapely>=2")
-    assert equals_exact(
-        from_wkb(geometry1), from_wkb(geometry2), tolerance=0.00001
+    assert shapely.equals_exact(
+        shapely.from_wkb(geometry1), shapely.from_wkb(geometry2), tolerance=0.00001
     ).all()
     assert all([np.array_equal(f1, f2) for f1, f2 in zip(field_data1, field_data2)])